+++ e2e/e2etests/test_solana_withdraw_and_call.go
@@ -62,6 +62,7 @@ func TestSolanaWithdrawAndCall(r *runner.E2ERunner, args []string) {
 			{PublicKey: [32]byte(r.ComputePdaAddress().Bytes()), IsWritable: false},
 			{PublicKey: [32]byte(r.GetSolanaPrivKey().PublicKey().Bytes()), IsWritable: true},
 			{PublicKey: [32]byte(solana.SystemProgramID.Bytes()), IsWritable: false},
+			{PublicKey: [32]byte(solana.SysVarInstructionsPubkey.Bytes()), IsWritable: false},
 		},
 		Data: []byte("hello"),
 	}