+++ e2e/e2etests/test_solana_withdraw_restricted_address.go
@@ -4,8 +4,11 @@ import (
 	"fmt"
 	"math/big"
 
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/gagliardetto/solana-go"
+	"github.com/gagliardetto/solana-go/rpc"
 	"github.com/stretchr/testify/require"
+	"github.com/zeta-chain/protocol-contracts/pkg/gatewayzevm.sol"
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
@@ -16,7 +19,8 @@ import (
 func TestSolanaWithdrawRestricted(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 2)
 
-	// parse restricted address
+	// ARRANGE
+	// Given amount, receiver, revert address
 	receiverRestricted, err := chains.DecodeSolanaWalletAddress(args[0])
 	require.NoError(r, err, fmt.Sprintf("unable to decode solana wallet address: %s", args[0]))
 
@@ -29,14 +33,50 @@ func TestSolanaWithdrawRestricted(r *runner.E2ERunner, args []string) {
 		withdrawAmount.Cmp(approvedAmount),
 		"Withdrawal amount must be less than the approved amount (1e9).",
 	)
+	revertAddress := r.EVMAddress()
 
+	// receiver balance before
+	result, err := r.SolanaClient.GetBalance(r.Ctx, receiverRestricted, rpc.CommitmentFinalized)
+	require.NoError(r, err)
+	receiverBalanceBefore := result.Value
+
+	// ACT
 	// withdraw
-	tx := r.WithdrawSOLZRC20(receiverRestricted, withdrawAmount, approvedAmount)
+	tx := r.WithdrawSOLZRC20(
+		receiverRestricted,
+		withdrawAmount,
+		approvedAmount,
+		gatewayzevm.RevertOptions{
+			RevertAddress:    revertAddress,
+			OnRevertGasLimit: big.NewInt(0),
+		},
+	)
 
-	// wait for the cctx to be mined
+	// wait for the withdraw tx to be mined
+	receipt := utils.MustWaitForTxReceipt(r.Ctx, r.ZEVMClient, tx, r.Logger, r.ReceiptTimeout)
+	utils.RequireTxSuccessful(r, receipt)
+
+	// revert address balance before
+	revertBalanceBefore, err := r.SOLZRC20.BalanceOf(&bind.CallOpts{}, revertAddress)
+	require.NoError(r, err)
+
+	// ASSERT
+	// wait for the cctx to be reverted
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
-	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_OutboundMined)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_Reverted)
 
-	// the cctx should be cancelled with zero value
-	r.VerifySolanaWithdrawalAmountFromCCTX(cctx, 0)
+	// the outbound should be cancelled with zero value
+	// note: the first outbound param is the cancel transaction
+	r.SolanaVerifyWithdrawalAmount(cctx.OutboundParams[0].Hash, 0)
+
+	// receiver balance should not change
+	result, err = r.SolanaClient.GetBalance(r.Ctx, receiverRestricted, rpc.CommitmentFinalized)
+	require.NoError(r, err)
+	receiverBalanceAfter := result.Value
+	require.EqualValues(r, receiverBalanceBefore, receiverBalanceAfter)
+
+	// revert address should receive the amount
+	revertBalanceAfter, err := r.SOLZRC20.BalanceOf(&bind.CallOpts{}, revertAddress)
+	require.NoError(r, err)
+	require.EqualValues(r, new(big.Int).Add(revertBalanceBefore, withdrawAmount), revertBalanceAfter)
 }