+++ zetaclient/chains/evm/signer/signer.go
@@ -24,7 +24,6 @@ import (
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/evm/client"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
-	"github.com/zeta-chain/node/zetaclient/compliance"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
 	"github.com/zeta-chain/node/zetaclient/logs"
 	"github.com/zeta-chain/node/zetaclient/zetacore"
@@ -341,19 +340,8 @@ func (signer *Signer) SignOutboundFromCCTX(
 	zetacoreClient interfaces.ZetacoreClient,
 	toChain zctx.Chain,
 ) (*ethtypes.Transaction, error) {
-	if compliance.IsCCTXRestricted(cctx) {
+	if !signer.PassesCompliance(cctx) {
 		// restricted cctx
-		compliance.PrintComplianceLog(
-			logger,
-			signer.Logger().Compliance,
-			true,
-			signer.Chain().ChainId,
-			cctx.Index,
-			cctx.InboundParams.Sender,
-			outboundData.to.Hex(),
-			cctx.GetCurrentOutboundParam().CoinType.String(),
-		)
-
 		return signer.SignCancel(ctx, outboundData)
 	} else if cctx.InboundParams.CoinType == coin.CoinType_Cmd {
 		// admin command
@@ -496,16 +484,17 @@ func (signer *Signer) BroadcastOutbound(
 
 	// define broadcast function
 	broadcast := func() error {
-		// get latest TSS account pending nonce
-		pendingNonce, err := signer.client.PendingNonceAt(ctx, signer.TSS().PubKey().AddressEVM())
+		// get latest TSS account nonce
+		latestNonce, err := signer.client.NonceAt(ctx, signer.TSS().PubKey().AddressEVM(), nil)
 		if err != nil {
 			return errors.Wrap(err, "unable to get latest TSS account nonce")
 		}
 
 		// if TSS nonce is higher than CCTX nonce, there is no need to broadcast
 		// this avoids foreseeable "nonce too low" error and unnecessary tracker report
-		if pendingNonce > nonce {
-			logger.Info().Uint64("tss_nonce", pendingNonce).Msg("cctx nonce is too low, skip broadcasting tx")
+		// Note: the latest finalized nonce is used here, not the pending nonce, making it possible to replacing pending txs
+		if latestNonce > nonce {
+			logger.Info().Uint64("latest_nonce", latestNonce).Msg("cctx nonce is too low, skip broadcasting tx")
 			return nil
 		}