+++ zetaclient/chains/solana/signer/execute.go
@@ -2,6 +2,7 @@ package signer
 
 import (
 	"context"
+	"fmt"
 
 	"cosmossdk.io/errors"
 	"github.com/ethereum/go-ethereum/common"
@@ -12,7 +13,6 @@ import (
 	"github.com/zeta-chain/node/pkg/chains"
 	contracts "github.com/zeta-chain/node/pkg/contracts/solana"
 	"github.com/zeta-chain/node/x/crosschain/types"
-	"github.com/zeta-chain/node/zetaclient/compliance"
 )
 
 // prepareExecuteTx prepares execute outbound
@@ -20,23 +20,10 @@ func (signer *Signer) prepareExecuteTx(
 	ctx context.Context,
 	cctx *types.CrossChainTx,
 	height uint64,
+	cancelTx bool,
 	logger zerolog.Logger,
 ) (outboundGetter, error) {
 	params := cctx.GetCurrentOutboundParam()
-	// compliance check
-	cancelTx := compliance.IsCCTXRestricted(cctx)
-	if cancelTx {
-		compliance.PrintComplianceLog(
-			logger,
-			signer.Logger().Compliance,
-			true,
-			signer.Chain().ChainId,
-			cctx.Index,
-			cctx.InboundParams.Sender,
-			params.Receiver,
-			"SOL",
-		)
-	}
 
 	// create msg execute
 	msg, msgIn, err := signer.createMsgExecute(cctx, cancelTx)
@@ -88,6 +75,12 @@ func (signer *Signer) createMsgExecute(
 		return nil, nil, errors.Wrapf(err, "cannot decode receiver address %s", params.Receiver)
 	}
 
+	// check sender based on execute type
+	sender, err := validateSender(cctx.InboundParams.Sender, executeType)
+	if err != nil {
+		return nil, nil, errors.Wrap(err, "cannot validate sender")
+	}
+
 	remainingAccounts := []*solana.AccountMeta{}
 	for _, a := range msg.Accounts {
 		remainingAccounts = append(remainingAccounts, &solana.AccountMeta{
@@ -101,7 +94,7 @@ func (signer *Signer) createMsgExecute(
 		nonce,
 		amount,
 		to,
-		cctx.InboundParams.Sender,
+		sender,
 		msg.Data,
 		executeType,
 		remainingAccounts,
@@ -171,4 +164,23 @@ func (signer *Signer) createExecuteInstruction(msg contracts.MsgExecute) (*solan
 	}
 
 	return inst, nil
+}
+
+// validateSender validates and formats the sender address based on execute type
+func validateSender(sender string, executeType contracts.ExecuteType) (string, error) {
+	if executeType == contracts.ExecuteTypeCall {
+		// for regular execute, sender should be an Ethereum address
+		senderEth := common.HexToAddress(sender)
+		if senderEth == (common.Address{}) {
+			return "", fmt.Errorf("invalid execute sender %s", sender)
+		}
+		return senderEth.Hex(), nil
+	}
+
+	// for revert execute, sender should be a Solana address
+	senderSol, err := solana.PublicKeyFromBase58(sender)
+	if err != nil {
+		return "", errors.Wrapf(err, "invalid execute revert sender %s", sender)
+	}
+	return senderSol.String(), nil
 }