+++ zetaclient/chains/solana/signer/signer.go
@@ -5,10 +5,10 @@ import (
 	"runtime/debug"
 	"time"
 
-	"cosmossdk.io/errors"
 	"github.com/gagliardetto/solana-go"
 	computebudget "github.com/gagliardetto/solana-go/programs/compute-budget"
 	"github.com/gagliardetto/solana-go/rpc"
+	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 
 	"github.com/zeta-chain/node/pkg/coin"
@@ -33,13 +33,24 @@ const (
 	// 6 retries will span over 1 + 2 + 4 + 8 + 16 + 32 + 64 = 127 seconds, good enough for the 2 minute timeout
 	broadcastRetries = 7
 
+	// pdaNonceWaitTimeout is the timeout for waiting for the PDA nonce to arrive
+	// given 1~2 seconds finality at the 'confirmed' level, 1 minute can cover 30~60 (the lookahead) parallel CCTXs processing
+	pdaNonceWaitTimeout = 1 * time.Minute
+
+	// broadcastOutboundCommitment is the commitment level for broadcasting solana outbound.
+	// Commitment "finalized" eliminate all risk but the tradeoff is pretty severe and effectively
+	// reduces the expiration of tx by about 13 seconds. The "confirmed" level has very low risk of
+	// belonging to a dropped fork.
+	// see: https://solana.com/developers/guides/advanced/confirmation#use-an-appropriate-preflight-commitment-level
+	broadcastOutboundCommitment = rpc.CommitmentConfirmed
+
 	// SolanaMaxComputeBudget is the max compute budget for a transaction.
 	SolanaMaxComputeBudget = 1_400_000
 )
 
 type Outbound struct {
-	Tx         *solana.Transaction
-	FallbackTx *solana.Transaction
+	Tx          *solana.Transaction
+	FallbackMsg *contracts.MsgIncrementNonce
 }
 
 type outboundGetter func() (*Outbound, error)
@@ -138,13 +149,14 @@ func (signer *Signer) TryProcessOutbound(
 		Str("cctx", cctx.Index).
 		Logger()
 
-	// support gas token only for Solana outbound
-	chainID := signer.Chain().ChainId
-	nonce := params.TssNonce
-	coinType := cctx.InboundParams.CoinType
-	isRevert := (cctx.CctxStatus.Status == types.CctxStatus_PendingRevert && cctx.RevertOptions.CallOnRevert)
-
-	var outboundGetter outboundGetter
+	var (
+		chainID        = signer.Chain().ChainId
+		nonce          = params.TssNonce
+		coinType       = cctx.InboundParams.CoinType
+		isRevert       = (cctx.CctxStatus.Status == types.CctxStatus_PendingRevert && cctx.RevertOptions.CallOnRevert)
+		cancelTx       = !signer.PassesCompliance(cctx)
+		outboundGetter outboundGetter
+	)
 
 	switch coinType {
 	case coin.CoinType_Cmd:
@@ -159,7 +171,7 @@ func (signer *Signer) TryProcessOutbound(
 	case coin.CoinType_Gas:
 		isRevert := (cctx.CctxStatus.Status == types.CctxStatus_PendingRevert && cctx.RevertOptions.CallOnRevert)
 		if cctx.IsWithdrawAndCall() || isRevert {
-			executeTxGetter, err := signer.prepareExecuteTx(ctx, cctx, height, logger)
+			executeTxGetter, err := signer.prepareExecuteTx(ctx, cctx, height, cancelTx, logger)
 			if err != nil {
 				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign execute outbound")
 				return
@@ -167,7 +179,7 @@ func (signer *Signer) TryProcessOutbound(
 
 			outboundGetter = executeTxGetter
 		} else {
-			withdrawTxGetter, err := signer.prepareWithdrawTx(ctx, cctx, height, logger)
+			withdrawTxGetter, err := signer.prepareWithdrawTx(ctx, cctx, height, cancelTx, logger)
 			if err != nil {
 				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign withdraw outbound")
 				return
@@ -178,7 +190,7 @@ func (signer *Signer) TryProcessOutbound(
 
 	case coin.CoinType_ERC20:
 		if cctx.IsWithdrawAndCall() || isRevert {
-			executeSPLTxGetter, err := signer.prepareExecuteSPLTx(ctx, cctx, height, logger)
+			executeSPLTxGetter, err := signer.prepareExecuteSPLTx(ctx, cctx, height, cancelTx, logger)
 			if err != nil {
 				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign execute spl outbound")
 				return
@@ -186,7 +198,7 @@ func (signer *Signer) TryProcessOutbound(
 
 			outboundGetter = executeSPLTxGetter
 		} else {
-			withdrawSPLTxGetter, err := signer.prepareWithdrawSPLTx(ctx, cctx, height, logger)
+			withdrawSPLTxGetter, err := signer.prepareWithdrawSPLTx(ctx, cctx, height, cancelTx, logger)
 			if err != nil {
 				logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign withdraw spl outbound")
 				return
@@ -195,7 +207,7 @@ func (signer *Signer) TryProcessOutbound(
 			outboundGetter = withdrawSPLTxGetter
 		}
 	case coin.CoinType_NoAssetCall:
-		executeTxGetter, err := signer.prepareExecuteTx(ctx, cctx, height, logger)
+		executeTxGetter, err := signer.prepareExecuteTx(ctx, cctx, height, cancelTx, logger)
 		if err != nil {
 			logger.Error().Err(err).Msgf("TryProcessOutbound: Fail to sign execute outbound")
 			return
@@ -217,6 +229,15 @@ func (signer *Signer) TryProcessOutbound(
 	// set relayer balance metrics
 	signer.SetRelayerBalanceMetrics(ctx)
 
+	// wait for the exact PDA nonce to arrive with timeout
+	ctxWait, cancel := context.WithTimeout(ctx, pdaNonceWaitTimeout)
+	defer cancel()
+
+	if err := signer.waitExactGatewayNonce(ctxWait, params.TssNonce); err != nil {
+		logger.Error().Err(err).Msgf("fail to wait for gateway nonce")
+		return
+	}
+
 	// Get transactions from getters
 	// This is when the recent block hash timer starts
 	outbound, err := outboundGetter()
@@ -236,7 +257,7 @@ func (signer *Signer) signTx(
 	limit uint64,
 ) (*solana.Transaction, error) {
 	// get a recent blockhash
-	recent, err := signer.client.GetLatestBlockhash(ctx, rpc.CommitmentFinalized)
+	recent, err := signer.client.GetLatestBlockhash(ctx, broadcastOutboundCommitment)
 	if err != nil {
 		return nil, errors.Wrap(err, "getLatestBlockhash error")
 	}
@@ -298,43 +319,54 @@ func (signer *Signer) broadcastOutbound(
 	// try broacasting tx with increasing backoff (1s, 2s, 4s, 8s, 16s, 32s, 64s)
 	// to tolerate tx nonce mismatch with PDA nonce or unknown RPC error
 	backOff := broadcastBackoff
-	for i := 0; i < broadcastRetries; i++ {
+	for range broadcastRetries {
 		time.Sleep(backOff)
 
-		// PDA nonce may already be increased by other relayer, no need to retry
-		pdaInfo, err := signer.client.GetAccountInfo(ctx, signer.pda)
+		// query the gateway PDA nonce
+		pdaNonce, err := signer.getGatewayNonce(ctx)
 		if err != nil {
-			logger.Error().Err(err).Fields(lf).Msgf("unable to get PDA account info")
-		} else {
-			pda, err := contracts.DeserializePdaInfo(pdaInfo)
-			if err != nil {
-				logger.Error().Err(err).Fields(lf).Msgf("unable to deserialize PDA info")
-			} else if pda.Nonce > nonce {
-				logger.Info().Err(err).Fields(lf).Msgf("PDA nonce %d is greater than outbound nonce, stop retrying", pda.Nonce)
-				break
-			}
-		}
-
-		// broadcast the signed tx to the Solana network with preflight check
-		txSig, err := signer.client.SendTransactionWithOpts(
-			ctx,
-			tx,
-			// Commitment "finalized" is too conservative for preflight check and
-			// it results in repeated broadcast attempts that only 1 will succeed.
-			// Commitment "processed" will simulate tx against more recent state
-			// thus fails faster once a tx is already broadcasted and processed by the cluster.
-			// This reduces the number of "failed" txs due to repeated broadcast attempts.
-			rpc.TransactionOpts{PreflightCommitment: rpc.CommitmentProcessed},
-		)
-		if err != nil {
-			if outbound.FallbackTx != nil && shouldUseFallbackTx(err, signer.GetGatewayAddress()) {
-				tx = outbound.FallbackTx
-			}
-			logger.Warn().Err(err).Fields(lf).Msgf("SendTransactionWithOpts failed")
+			logger.Error().Err(err).Fields(lf).Msg("unable to get PDA nonce")
 			backOff *= 2
 			continue
 		}
-		logger.Info().Fields(lf).Msgf("broadcasted Solana outbound successfully")
+		lf["pda_nonce"] = pdaNonce
+
+		// PDA nonce may already be increased by other relayers, no need to retry
+		if pdaNonce > nonce {
+			logger.Info().Fields(lf).Msg("PDA nonce is greater than outbound nonce, stop retrying")
+			break
+		}
+
+		// broadcast the signed tx to the Solana network with preflight check
+		// the PDA nonce MUST be equal to 'nonce' if arrived here, guaranteed by upstream code
+		txSig, err := signer.client.SendTransactionWithOpts(
+			ctx,
+			tx,
+			rpc.TransactionOpts{PreflightCommitment: broadcastOutboundCommitment},
+		)
+		if err != nil {
+			shouldUseFallbackTx, failureReason := parseRPCErrorForFallback(err, signer.GetGatewayAddress())
+			if outbound.FallbackMsg != nil && shouldUseFallbackTx {
+				// create and sign fallback transaction
+				outbound.FallbackMsg.SetFailureReason(failureReason)
+				fallbackInst, err := signer.createIncrementNonceInstruction(*outbound.FallbackMsg)
+				if err != nil {
+					logger.Error().Err(err).Fields(lf).Msg("error creating increment nonce instruction")
+					break
+				}
+
+				fallbackTx, err := signer.signTx(ctx, fallbackInst, 0)
+				if err != nil {
+					logger.Error().Err(err).Fields(lf).Msg("error signing increment nonce instruction")
+					break
+				}
+				tx = fallbackTx
+			}
+			logger.Warn().Err(err).Fields(lf).Msg("SendTransactionWithOpts failed")
+			backOff *= 2
+			continue
+		}
+		logger.Info().Fields(lf).Msg("broadcasted Solana outbound successfully")
 
 		// successful broadcast; report to the outbound tracker
 		signer.reportToOutboundTracker(ctx, zetacoreClient, chainID, nonce, txSig, logger)
@@ -356,20 +388,9 @@ func (signer *Signer) createOutboundWithFallback(
 		return nil, errors.Wrap(err, "error signing main instruction")
 	}
 
-	// Create and sign fallback transaction
-	fallbackInst, err := signer.createIncrementNonceInstruction(*msgIn)
-	if err != nil {
-		return nil, errors.Wrap(err, "error creating increment nonce instruction")
-	}
-
-	fallbackTx, err := signer.signTx(ctx, fallbackInst, 0)
-	if err != nil {
-		return nil, errors.Wrap(err, "error signing fallback instruction")
-	}
-
 	return &Outbound{
-		Tx:         tx,
-		FallbackTx: fallbackTx,
+		Tx:          tx,
+		FallbackMsg: msgIn,
 	}, nil
 }
 
@@ -457,4 +478,77 @@ func signMsgWithFallback[T SignableMessage[T]](
 	}
 
 	return msg.SetSignature(signature[0]), msgIn.SetSignature(signature[1]), nil
+}
+
+// waitExactGatewayNonce waits for exact given gateway nonce to arrive
+//
+// the reasons are:
+//  1. any pre-signed Solana tx expires after 150 blocks, so we should avoid pre-signing any tx to maximize the lifetime of signed txs
+//  2. there can be up to 'lookahead' CCTX processing goroutines running in parallel, so waiting for PDA nonce helps to order the CCTX
+//     processing goroutines by nonce and avoid nonce mismatch
+//  3. less nonce mismatch will reduce CCTX retries and TSS keysign requests
+func (signer *Signer) waitExactGatewayNonce(ctx context.Context, nonce uint64) error {
+	logger := signer.Logger().Std.With().
+		Str("method", "waitExactGatewayNonce").
+		Int64("chain", signer.Chain().ChainId).
+		Uint64("nonce", nonce).
+		Logger()
+
+	for {
+		if ctx.Err() != nil {
+			return errors.Wrap(ctx.Err(), "context cancelled")
+		}
+
+		// check timeout to avoid infinite waiting
+		if deadline, ok := ctx.Deadline(); ok {
+			if time.Now().After(deadline) {
+				return errors.New("timeout reached on waiting for gateway nonce")
+			}
+		}
+
+		// query the gateway PDA nonce
+		pdaNonce, err := signer.getGatewayNonce(ctx)
+		if err != nil {
+			logger.Error().Err(err).Msgf("unable to get gateway nonce")
+			time.Sleep(time.Second) // prevent RPC spamming
+			continue
+		}
+
+		switch {
+		case pdaNonce > nonce:
+			return errors.Wrapf(err, "PDA nonce %d is greater than outbound nonce %d", pdaNonce, nonce)
+		case pdaNonce == nonce:
+			return nil
+		default:
+			logger.Info().Uint64("pda.nonce", pdaNonce).Msg("waiting for PDA nonce to arrive")
+
+			// calculate how far behind the PDA nonce and sleep accordingly
+			//  - base sleep time of 1 second, multiplied by the nonce difference
+			//  - 'lookahead' parameter should keep this from getting too out of control
+			// #nosec G115 always in range
+			sleepDuration := time.Second * time.Duration(nonce-pdaNonce)
+			time.Sleep(sleepDuration)
+		}
+	}
+}
+
+// getGatewayNonce queries the gateway nonce from the PDA account information
+func (signer *Signer) getGatewayNonce(ctx context.Context) (uint64, error) {
+	// query the gateway PDA account information
+	pdaInfo, err := signer.client.GetAccountInfoWithOpts(
+		ctx,
+		signer.pda,
+		&rpc.GetAccountInfoOpts{Commitment: broadcastOutboundCommitment},
+	)
+	if err != nil {
+		return 0, errors.Wrap(err, "unable to get gateway PDA account info")
+	}
+
+	// deserialize the PDA account information
+	pda, err := contracts.DeserializePdaInfo(pdaInfo)
+	if err != nil {
+		return 0, errors.Wrap(err, "unable to deserialize PDA info")
+	}
+
+	return pda.Nonce, nil
 }