+++ zetaclient/chains/sui/signer/signer.go
@@ -14,7 +14,6 @@ import (
 	cctypes "github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
-	"github.com/zeta-chain/node/zetaclient/compliance"
 	"github.com/zeta-chain/node/zetaclient/logs"
 )
 
@@ -106,10 +105,20 @@ func (s *Signer) ProcessCCTX(ctx context.Context, cctx *cctypes.CrossChainTx, ze
 		return errors.Wrap(err, "unable to create cancel tx builder")
 	}
 
-	var txDigest string
+	var (
+		txDigest      string
+		validReceiver = true
+	)
+
+	// check CCTX receiver address format
+	receiver := cctx.GetCurrentOutboundParam().Receiver
+	if err := sui.ValidateAddress(receiver); err != nil {
+		validReceiver = false
+		logger.Error().Err(err).Str("receiver", receiver).Msg("Invalid receiver address")
+	}
 
 	// broadcast tx according to compliance check result
-	if s.passesCompliance(cctx) {
+	if validReceiver && s.PassesCompliance(cctx) {
 		txDigest, err = s.broadcastWithdrawalWithFallback(ctx, withdrawTxBuilder, cancelTxBuilder)
 	} else {
 		txDigest, err = s.broadcastCancelTx(ctx, cancelTxBuilder)
@@ -200,28 +209,6 @@ func (s *Signer) SignTxWithCancel(
 	}
 
 	return sig, sigCancel, nil
-}
-
-func (s *Signer) passesCompliance(cctx *cctypes.CrossChainTx) bool {
-	restricted := compliance.IsCCTXRestricted(cctx)
-	if !restricted {
-		return true
-	}
-
-	params := cctx.GetCurrentOutboundParam()
-
-	compliance.PrintComplianceLog(
-		s.Logger().Std,
-		s.Logger().Compliance,
-		true,
-		s.Chain().ChainId,
-		cctx.Index,
-		cctx.InboundParams.Sender,
-		params.Receiver,
-		params.CoinType.String(),
-	)
-
-	return false
 }
 
 // wrapDigest wraps the digest with sha256.