+++ changelog.md
@@ -1,6 +1,24 @@
 # CHANGELOG
 
-# UNRELEASED
+## UNRELEASED
+
+### Features
+
+* [3929](https://github.com/zeta-chain/node/pull/3929) - add support for TON http-rpc
+
+### Fixes
+
+* [3953](https://github.com/zeta-chain/node/pull/3953) - skip Bitcoin outbound when scanning inbound transactions
+* [3957](https://github.com/zeta-chain/node/pull/3957) - remove tx format assumption from solana parse gateway instruction
+* [3956](https://github.com/zeta-chain/node/pull/3956) - use the latest nonce to perform pre-broadcast check to make evm tx replacement possible
+* [3954](https://github.com/zeta-chain/node/pull/3954) - fail sui withdrawal event in the ZEVM if it carries invalid receiver address
+* [3917](https://github.com/zeta-chain/node/pull/3917) - prevent jailed observers from voting
+
+### Refactor
+
+* [3940](https://github.com/zeta-chain/node/pull/3940) - avoid pre-signing solana outbound by waiting for the exact PDA nonce to arrive
+
+## v31.0.0
 
 ### Breaking Changes
 
@@ -28,13 +46,16 @@
 * [3396](https://github.com/zeta-chain/node/pull/3396) - add support for Bitcoin RBF (Replace-By-Fee) in zetaclient
 * [3864](https://github.com/zeta-chain/node/pull/3864) - add compliance checks for TON inbounds
 * [3881](https://github.com/zeta-chain/node/pull/3881) - add zetatool cmd to analyze size of application.db
+* [3906](https://github.com/zeta-chain/node/pull/3906) - revert restricted cctx for EVM, bitcoin and solana chains
+* [3918](https://github.com/zeta-chain/node/pull/3918) - attach failure reason to solana increment nonce
+* [3932](https://github.com/zeta-chain/node/pull/3932) - hardcoded block time related params
 
 ### Refactor
 
 * [3709](https://github.com/zeta-chain/node/pull/3709) - improve cctx error message for out of gas errors when creating outbound
 * [3777](https://github.com/zeta-chain/node/pull/3777) - use SignBatch keysign for solana outbound tx and fallback tx
 * [3813](https://github.com/zeta-chain/node/pull/3813) - set ZETA protocol fee to 0
-* [3848g](https://github.com/zeta-chain/node/pull/3848) - extend min gas limit check to prevent intrinsic low gas limit
+* [3848](https://github.com/zeta-chain/node/pull/3848) - extend min gas limit check to prevent intrinsic low gas limit
 
 ### Fixes
 
@@ -51,7 +72,13 @@
 * [3890](https://github.com/zeta-chain/node/pull/3890) - solana abort address format
 * [3901](https://github.com/zeta-chain/node/pull/3901) - prevent cctx being set as abortRefunded if the abort processing failed before the refund
 * [3872](https://github.com/zeta-chain/node/pull/3872) - delete testnet ballots for creation height 0 and add a query to list all ballots created at a height.
-* [3917](https://github.com/zeta-chain/node/pull/3917) - prevent jailed observers from voting
+* [3916](https://github.com/zeta-chain/node/pull/3916) - infinite scan in filter solana inbound events
+* [3914](https://github.com/zeta-chain/node/pull/3914) - check tx result err in filter inbound events
+* [3904](https://github.com/zeta-chain/node/pull/3904) - improve observer emissions distribution to maximise pool utilisation
+* [3895](https://github.com/zeta-chain/node/pull/3895) - solana call required accounts number condition
+* [3896](https://github.com/zeta-chain/node/pull/3896) - add sender to solana execute message hash
+* [3920](https://github.com/zeta-chain/node/pull/3920) - show correct gas limit for synthetic txs
+* [3934](https://github.com/zeta-chain/node/pull/3934) - post zero priority fee for EVM chains to avoid gas price pump failure in the zetacore
 
 ### Tests
 
@@ -60,6 +87,8 @@
 * [3582](https://github.com/zeta-chain/node/pull/3852) - add solana to tss migration e2e tests
 * [3866](https://github.com/zeta-chain/node/pull/3866) - add e2e test for upgrading sui gateway package
 * [3417](https://github.com/zeta-chain/node/pull/3417) - add e2e test for the Bitcoin RBF (Replace-By-Fee) feature
+* [3885](https://github.com/zeta-chain/node/pull/3885) - add e2e test for MsgAddObserver
+* [3893](https://github.com/zeta-chain/node/pull/3893) - add e2e performance tests for sui deposit and withdrawal
 
 
 ### Refactor